refactor(comic_builder): modularize into state/, generation/, and api/ submodules
- Extract domain types from consolidated_state.py into state/ submodule
  (base.py, artifact.py, character.py, panel.py, story.py, consolidated.py)
- Group pipeline services into generation/ submodule
  (character_extractor, character_renderer, panel_generator)
- Move FastAPI dependencies to api/dependencies.py where they belong
- Rename project_state_manager.py → state_manager.py
- Centralize RenderError in exceptions.py

Improves discoverability: domain types in state/, business logic in generation/,
HTTP concerns in api/. Clean re-exports via __init__.py preserve import ergonomics.
