+++ backend/core/comic_builder/api/dependencies.py
@@ -9,7 +9,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from core.auth.dependencies import get_current_user_id
 from core.services.database import get_async_db_session
 
-from .models import Project
+from ..models import Project
 
 router = APIRouter(prefix="/phase", tags=["comic", "builder"])
 
@@ -35,3 +35,4 @@ async def verify_project_access(
         )
 
     return project.id
+
