+++ backend/core/comic_builder/generation/__init__.py
@@ -1,0 +1,10 @@
+from .character_extractor import CharacterExtractor
+from .character_renderer import CharacterRenderer
+from .panel_generator import PanelGenerator
+
+__all__ = [
+    "CharacterExtractor",
+    "CharacterRenderer",
+    "PanelGenerator",
+]
+
