+++ backend/core/comic_builder/state/artifact.py
@@ -1,0 +1,5 @@
+from .base import BaseComicStateEntity
+
+
+class Artifact(BaseComicStateEntity):
+    url: str | None = None
