+++ backend/core/comic_builder/state/character.py
@@ -1,25 +1,11 @@
-import uuid
-from typing import Any, Literal
+from typing import Literal
 
 from pydantic import Field
 
 from core.common import AliasedBaseModel
 
-StreamingStatus = Literal["idle", "streaming", "completed", "error"]
-
-
-class BaseComicStateEntity(AliasedBaseModel):
-    id: uuid.UUID = Field(default_factory=uuid.uuid4)
-    user_input_text: list[str] = Field(default_factory=list)
-    status: StreamingStatus = Field(default="idle")
-
-
-class Artifact(BaseComicStateEntity):
-    url: str | None = None
-
-
-class Story(BaseComicStateEntity):
-    story_text: str = Field(default="")
+from .artifact import Artifact
+from .base import BaseComicStateEntity
 
 
 class CharacterBase(AliasedBaseModel):
@@ -107,33 +93,3 @@ class CharacterBase(AliasedBaseModel):
 class Character(BaseComicStateEntity, CharacterBase):
     render: Artifact | None = None
 
-
-class ComicPanelBase(BaseComicStateEntity):
-    background: str = Field(
-        ...,
-        description="The background of the panel that will be rendered as a background image.",
-    )
-    characters: list[str] = Field(
-        ..., description="The names of the characters in the panel."
-    )
-    dialogue: str = Field(
-        ...,
-        description="The dialogue in the panel that will be rendered as speech bubbles.",
-    )
-
-class ComicPanel(ComicPanelBase):
-    render: Artifact | None = None
-
-
-class ConsolidatedComicState(AliasedBaseModel):
-    story: Story
-    characters: dict[uuid.UUID, Character] = Field(default_factory=dict)
-    panels: list[ComicPanel] = Field(default_factory=list)
-
-
-def initialize_empty_consolidated_state_dict() -> dict[str, Any]:
-    return ConsolidatedComicState(
-        story=Story(),
-        characters={},
-        panels=[],
-    ).model_dump(by_alias=True)
