+++ backend/core/comic_builder/state/story.py
@@ -1,0 +1,8 @@
+from pydantic import Field
+
+from .base import BaseComicStateEntity
+
+
+class Story(BaseComicStateEntity):
+    story_text: str = Field(default="")
+
